+++ mobile/android/android-components/docs/changelog.md
@@ -24,6 +24,9 @@ permalink: /changelog/
 * **support-migration**
   * ⚠️ **This is a breaking change**: `FennecMigrator` now takes `Lazy` references to storage layers.
 
+* **concept-storage**, **service-sync-logins**
+  * 🆕 New API: `PlacesStorage#warmUp`, `SyncableLoginsStorage#warmUp` - allows consumers to ensure that underlying storage database connections are fully established.
+
 # 37.0.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v36.0.0...v37.0.0)